# platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CASS API
  version: 1.0.0
  description: The control-plane API of CASS, the Catastrophe Analytics and Scenario
    Suite. The React application calls only this API; it never calls OpenQuake or
    Oasis directly.
paths:
...
  /api/v1/artifacts/:
    get:
      operationId: api_v1_artifacts_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: project
        schema:
          type: string
          format: uuid
      - in: query
        name: retention
        schema:
          type: string
          enum:
          - diagnostic
          - hazard_intermediate
          - model_asset
          - operational
          - permanent
          - portfolio
          - result
        description: |-
          * `permanent` - Permanent
          * `model_asset` - Model Asset
          * `hazard_intermediate` - Hazard Intermediate
          * `portfolio` - Portfolio
          * `result` - Result
          * `diagnostic` - Diagnostic
          * `operational` - Operational
      - in: query
        name: role
        schema:
          type: string
      - in: query
        name: state
        schema:
          type: string
          enum:
          - expired
          - pending
          - quarantined
          - registered
        description: |-
          * `pending` - Awaiting upload
          * `registered` - Registered
          * `quarantined` - Quarantined
          * `expired` - Expired
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedArtifactList'
          description: ''
  /api/v1/artifacts/{id}/:
    get:
      operationId: api_v1_artifacts_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this artifact.
        required: true
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
          description: ''
...
  /api/v1/artifacts/upload-session/:
    post:
      operationId: api_v1_artifacts_upload_session_create
      description: |-
        Issue a short-lived direct-to-store upload grant.

        Section 5 has the browser upload large files straight to object
        storage; Django registers the artifact only once the bytes are present
        and checksummed.
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
          description: ''
...
  /api/v1/runs/{id}/artifacts/:
    get:
      operationId: api_v1_runs_artifacts_retrieve
      description: Inputs and outputs recorded against this run.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this run.
        required: true
      tags:
      - api
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
          description: ''
...
components:
  schemas:
    AccessEnum:
      enum:
      - project
      - model
      - platform
      type: string
      description: |-
        * `project` - Project
        * `model` - Model
        * `platform` - Platform
...
    Artifact:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        uri:
          type: string
          readOnly: true
        checksum:
          type: string
          readOnly: true
          description: Algorithm-prefixed content digest, recorded at registration.
        size_bytes:
          type: integer
          readOnly: true
        content_type:
          type: string
          readOnly: true
        retention:
          allOf:
          - $ref: '#/components/schemas/RetentionEnum'
          readOnly: true
        access:
          allOf:
          - $ref: '#/components/schemas/AccessEnum'
          readOnly: true
        state:
          allOf:
          - $ref: '#/components/schemas/ArtifactStateEnum'
          readOnly: true
        project:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        role:
          type: string
          readOnly: true
          description: What this artifact is in its manifest, such as oed_location
            or footprint.
        original_filename:
          type: string
          readOnly: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        readable:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - access
      - checksum
      - content_type
      - created_at
      - expires_at
      - id
      - original_filename
      - project
      - readable
      - retention
      - role
      - size_bytes
      - state
      - uri
    ArtifactStateEnum:
      enum:
      - pending
      - registered
      - quarantined
      - expired
      type: string
      description: |-
        * `pending` - Awaiting upload
        * `registered` - Registered
        * `quarantined` - Quarantined
        * `expired` - Expired
...
    KindEnum:
      enum:
      - hazard
      - conversion
      - analysis
      type: string
      description: |-
        * `hazard` - Hazard run
        * `conversion` - Conversion run
        * `analysis` - Analysis run
...
    PaginatedArtifactList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
...
    RetentionEnum:
      enum:
      - permanent
      - model_asset
      - hazard_intermediate
      - portfolio
      - result
      - diagnostic
      - operational
      type: string
      description: |-
        * `permanent` - Permanent
        * `model_asset` - Model Asset
        * `hazard_intermediate` - Hazard Intermediate
        * `portfolio` - Portfolio
        * `result` - Result
        * `diagnostic` - Diagnostic
        * `operational` - Operational
    Run:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        kind:
          allOf:
          - $ref: '#/components/schemas/KindEnum'
          readOnly: true
        project:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: Null for model-build runs, which belong to a model version
            rather than a project.
        label:
          type: string
          readOnly: true
        state:
          allOf:
          - $ref: '#/components/schemas/RunStateEnum'
          readOnly: true
        stage:
          type: string
          readOnly: true
          description: Current pipeline stage key.
        stage_label:
          type: string
          readOnly: true
        progress:
          type: number
          format: double
          readOnly: true
        pipeline:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
        execution_profile:
          type: string
          readOnly: true
          description: Declared CPU, memory and timeout envelope from section 11.
        correlation_id:
          type: string
          readOnly: true
        queued_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        started_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        finished_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        duration_seconds:
          type: integer
          readOnly: true
          nullable: true
        peak_memory_mb:
          type: integer
          readOnly: true
          nullable: true
        failure_stage:
          type: string
          readOnly: true
        failure_summary:
          type: string
          readOnly: true
        failure_detail:
          type: string
          readOnly: true
        settings_hash:
          type: string
          readOnly: true
        retry_of:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        may_retry:
          type: boolean
          readOnly: true
        may_publish_results:
          type: boolean
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - correlation_id
      - created_at
      - duration_seconds
      - execution_profile
      - failure_detail
      - failure_stage
      - failure_summary
      - finished_at
      - id
      - is_active
      - kind
      - label
      - may_publish_results
      - may_retry
      - peak_memory_mb
      - pipeline
      - progress
      - project
      - queued_at
      - retry_of
      - settings_hash
      - stage
      - stage_label
      - started_at
      - state
    RunStateEnum:
      enum:
      - draft
      - queued
      - running
      - blocked
      - cancelling
      - cancelled
      - failed
      - succeeded
      type: string
      description: |-
        * `draft` - Draft
        * `queued` - Queued
        * `running` - Running
        * `blocked` - Blocked
        * `cancelling` - Cancelling
        * `cancelled` - Cancelled
        * `failed` - Failed
        * `succeeded` - Succeeded
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"